+++ CHANGELOG.rst
@@ -6,7 +6,17 @@ All notable changes to this project will be documented in this file. Dates are i
 Unreleased
 ==========
 
-[0.4.1] - 2017-06-13
+[0.4.2] - 2018-03-16
+====================
+
+Added
+-----
+- Support get analysis by project.
+- Support download anaylsis.
+- Support set status of project.
+- Support set status of job.
+
+[0.4.1] - 2018-01-09
 ====================
 
 Added